+++ src/app/components/layouts/portlet/terra-portlet.component.spec.ts
@@ -65,10 +65,11 @@ describe('TerraPortletComponent', () =>
         expect(component.inputPortletHeader).toBeUndefined();
         expect(component.inputIsCollapsable).toBe(false);
         expect(component.inputCollapsed).toBe(false);
+        expect(component.inputIsDisabled).toBe(false);
         expect(component.inputButtonList).toEqual([]);
     });
 
-    it(`should set classes accordingly to 'inputIsCollapsable' and 'inputHighlightPortlet'`, () =>
+    it(`should set classes accordingly to 'inputIsCollapsable', 'inputHighlightPortlet' and 'inputIsDisabled'`, () =>
     {
         const portletDiv:DebugElement = debugElement.query(By.css('div.portlet'));
 
@@ -79,6 +80,10 @@ describe('TerraPortletComponent', () =>
         component.inputHighlightPortlet = true;
         fixture.detectChanges();
         expect(portletDiv.classes['highlight']).toBe(true);
+
+        component.inputIsDisabled = true;
+        fixture.detectChanges();
+        expect(portletDiv.classes['disabled']).toBe(true);
     });
 
     it(`should have a visible header after 'inputPortletHeader' is set`, () =>
@@ -134,6 +139,28 @@ describe('TerraPortletComponent', () =>
 
         expect(debugElement.query(By.css('span.icon-collapse_down'))).toBeFalsy();
         expect(debugElement.query(By.css('span.icon-collapse_up'))).toBeTruthy();
+    });
+
+    it(`should keep the collapse state when disabled`, () =>
+    {
+        component.inputPortletHeader = portletHeader;
+        component.inputIsCollapsable = true;
+        component.inputIsDisabled = true;
+        fixture.detectChanges();
+
+        let onHeaderClick:Spy = spyOn(component, 'toggleCollapse');
+
+        expect(component.inputCollapsed).toBe(true);
+        expect(debugElement.query(By.css('span.icon-collapse_down'))).toBeTruthy();
+        expect(debugElement.query(By.css('span.icon-collapse_up'))).toBeFalsy();
+
+        debugElement.query(By.css('div.portlet-head')).triggerEventHandler('click', null);
+        fixture.detectChanges();
+
+        expect(onHeaderClick).toHaveBeenCalled();
+        expect(component.inputCollapsed).toBe(true);
+        expect(debugElement.query(By.css('span.icon-collapse_down'))).toBeTruthy();
+        expect(debugElement.query(By.css('span.icon-collapse_up'))).toBeFalsy();
     });
 
     it(`should collapse icon only of 'inputIsCollapsable' set to true`, () =>